[modify] state 위치변경
