+++ src/components/CalendarPage/CalendarWapper.tsx
@@ -7,9 +7,8 @@ import CalendarTileContent from '@/components/CalendarPage/CalendarTileContent';
 import { dateformat } from '@/utils/date';
 import 'react-calendar/dist/Calendar.css'; // css import
 
-const CalendarWapper = () => {
+const CalendarWapper = ({ setValue, value, diaryList }: any) => {
   const [showModal, setShowModal] = useState<boolean>(false);
-  const [value, onChange] = useState<Date>(new Date());
 
   const diaryDate = dateformat(value);
   const time = value.getTime();
@@ -18,53 +17,10 @@ const CalendarWapper = () => {
     setShowModal(false);
   }, [time]);
 
-  const diaryList = [
-    {
-      id: 1,
-      date: '2023-02-24',
-      contents: {
-        text: '즐거웠다',
-        photo:
-          'https://item.kakaocdn.net/do/aebede13eed766c14f8e46d68509586c7154249a3890514a43687a85e6b6cc82',
-        emotion: '슬픔',
-      },
-    },
-    {
-      id: 2,
-      date: '2023-02-25',
-      contents: {
-        text: '',
-        photo:
-          'https://item.kakaocdn.net/do/aebede13eed766c14f8e46d68509586c7154249a3890514a43687a85e6b6cc82',
-        emotion: '슬픔',
-      },
-    },
-    {
-      id: 3,
-      date: '2023-02-26',
-      contents: {
-        text: '춤을 췄다.',
-        photo:
-          'https://item.kakaocdn.net/do/aebede13eed766c14f8e46d68509586c7154249a3890514a43687a85e6b6cc82',
-        emotion: '쏘쏘',
-      },
-    },
-    {
-      id: 4,
-      date: '2023-02-26',
-      contents: {
-        text: '춤을 췄다.',
-        photo:
-          'https://item.kakaocdn.net/do/aebede13eed766c14f8e46d68509586c7154249a3890514a43687a85e6b6cc82',
-        emotion: '쏘쏘',
-      },
-    },
-  ];
-
   return (
     <CalendarContainer>
       <Calendar
-        onChange={onChange}
+        onChange={setValue}
         next2Label={null}
         prev2Label={null}
         prevLabel={<img src="/left_arrow.svg" width={20} height={20} />}
@@ -74,7 +30,7 @@ const CalendarWapper = () => {
         formatDay={(locale, date) => moment(date).format('D')} // '일' 표시 x
         tileContent={({ date }) => {
           const dateFormated = dateformat(date);
-          const listFinded = diaryList.find((diary) => diary.date === dateFormated);
+          const listFinded = diaryList.find((diary: any) => diary.date === dateFormated);
 
           if (listFinded) {
             return (
@@ -90,21 +46,6 @@ const CalendarWapper = () => {
           }
         }}
       />
-      <div>{diaryDate}</div>
-      {diaryList
-        .filter((diary) => diary.date === diaryDate)
-        .map((item) => (
-          <div key={item.id}>
-            <div>{item.contents.text}</div>
-            <img
-              src={item.contents.photo}
-              style={{
-                width: 250,
-                height: 250,
-              }}
-            />
-          </div>
-        ))}
     </CalendarContainer>
   );
 };
@@ -112,6 +53,7 @@ const CalendarWapper = () => {
 const CalendarContainer = styled.div`
   .react-calendar {
     background-color: transparent;
+    border: none;
   }
 
   .react-calendar__navigation {
