+++ src/pages/CalendarPage.tsx
@@ -1,6 +1,54 @@
+import { useState } from 'react';
+
 import CalendarWapper from '@/components/CalendarPage/CalendarWapper';
+import Diary from '@/components/CalendarPage/Diary';
 
 const CalendarPage = () => {
+  const [value, setValue] = useState<Date>(new Date());
+
+  const diaryList = [
+    {
+      id: 1,
+      date: '2023-02-24',
+      contents: {
+        text: '즐거웠다',
+        photo:
+          'https://item.kakaocdn.net/do/aebede13eed766c14f8e46d68509586c7154249a3890514a43687a85e6b6cc82',
+        emotion: '슬픔',
+      },
+    },
+    {
+      id: 2,
+      date: '2023-02-25',
+      contents: {
+        text: '',
+        photo:
+          'https://item.kakaocdn.net/do/aebede13eed766c14f8e46d68509586c7154249a3890514a43687a85e6b6cc82',
+        emotion: '슬픔',
+      },
+    },
+    {
+      id: 3,
+      date: '2023-02-26',
+      contents: {
+        text: '춤을 췄다.',
+        photo:
+          'https://item.kakaocdn.net/do/aebede13eed766c14f8e46d68509586c7154249a3890514a43687a85e6b6cc82',
+        emotion: '쏘쏘',
+      },
+    },
+    {
+      id: 4,
+      date: '2023-02-26',
+      contents: {
+        text: '춤을 췄다.',
+        photo:
+          'https://item.kakaocdn.net/do/aebede13eed766c14f8e46d68509586c7154249a3890514a43687a85e6b6cc82',
+        emotion: '쏘쏘',
+      },
+    },
+  ];
+
   return (
     <div
       style={{
@@ -39,7 +87,8 @@ const CalendarPage = () => {
           통계
         </div>
       </div>
-      <CalendarWapper />
+      <CalendarWapper setValue={setValue} value={value} diaryList={diaryList} />
+      <Diary value={value} diaryList={diaryList} />
     </div>
   );
 };
